+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-# [Unreleased]
+# 1.3.3 (2017-04-03)
  - Fix InfluxDB response / header
+ - Socket and HTTP timeout parameters
  
 # 1.3.2 (2017-03-24)
 - Remove jars from build